make compilation quiet.
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@2180 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/include/asm-blackfin/uaccess.h
@@ -171,7 +171,7 @@ static inline long copy_from_user(void *to, const void __user *from, unsigned lo
 static inline long strncpy_from_user(char *dst, const char *src, long count)
 {
 	char *tmp;
-	if ((unsigned long)src > memory_end || ((unsigned long)src < _stext)) {
+	if ((unsigned long)src > (unsigned long)memory_end || ((unsigned long)src < (unsigned long)_stext)) {
 		return -EFAULT;
 	}
 	strncpy(dst, src, count);
